RouteParseTests: Write tests for `parse()` taking `Path`
Also move parsing-related tests to `RouteParseTests`.
