+++ shared/src/test/scala/trail/RouteExampleTests.scala
@@ -5,23 +5,14 @@ import org.scalatest._
 import shapeless.HNil
 
 class RouteExampleTests extends FlatSpec with Matchers {
-  "A simple example" should "just work" in {
-    val userInfo     = Root / "user" / Arg[String] / Arg[Boolean]
-    val userInfoUrl  = userInfo("bob" :: false :: HNil)
-    val userInfo2    = (Root / "user" / "bob" / false).url(HNil)
+  "url()" should "work with mandatory arguments" in {
+    val userInfo  = Root / "user" / Arg[String] / Arg[Boolean]
 
-    assert(userInfoUrl === "/user/bob/false")
-    assert(userInfoUrl === userInfo2)
+    val url1 = userInfo("bob" :: false :: HNil)
+    val url2 = (Root / "user" / "bob" / false).url(HNil)
 
-    val parsed = userInfo.parse("/user/bob/true")
-    parsed shouldBe defined
-
-    assert(parsed === Some("bob" :: true :: HNil))
-
-    userInfo.parse("/user/bob") shouldBe empty
-    userInfo.parse("/user/bob/true/true") shouldBe empty
-    userInfo.parse("/user/bob/1") shouldBe empty
-    userInfo.parse("/usr/bob/1") shouldBe empty
+    assert(url1 === "/user/bob/false")
+    assert(url2 === "/user/bob/false")
   }
 
   "url()" should "work" in {
@@ -36,13 +27,6 @@ class RouteExampleTests extends FlatSpec with Matchers {
     val url  = list(HNil, Option.empty[Int] :: Option(true) :: HNil)
 
     assert(url == "/list?upload=true")
-  }
-
-  "parse()" should "work" in {
-    val userInfo = Root / "user" / Arg[String] / Arg[Boolean]
-
-    assert(userInfo.parse("/user/hello/false")
-      .contains("hello" :: false :: HNil))
   }
 
   "Expressing a routing table" should "be possible with pattern matching" in {
